Add bucket property `node_confirms` for physical diversity
Briefly, some Riak deployments require that a write is on multiple
physical nodes before they're safe.

Given a good ring, using Primaries is one way to ensure this physical
diversity is to use `pw=2` (for example) to ensure two physical nodes
for the write. This has an availability cost. With 5 nodes and 2 nodes
down storing on 2 physical nodes is still very possible, but 2
primaries will leave a range of preflists unavailable to write. This
commit adds a bucket property and put option of `node_confrims` to the
declare the number of diverse physical node acks required for a write
to be successful. For the above example `node_confirms=2` would be
used, and as long as any two diverse nodes (primary OR fallback)
reply, the write is a success.