Logging
The aae-status is not a good way of recording the exchange activity - would prefer to have information on strating exchanges in logs to observe trends through log analysis.

Hypothesis is that exchanges aren't that common as to worry about the additional log for each exchange where  a start attempt was made.