Fix intermittent failure of `hll` bucket prop
The `gen_valid_bucket` call used regular `gen_hll_precision` meaning
that every now and again the _valid_ bucket had an _invalid_ hll
property. When this coincides with the update property also having the
same invalid hll property the resulting bucket has an invalid hll
property that matches the input and so it appears that validation
failed, and rightly the test fails. However validation worked, the
original bucket was invalid is all. This commit adds a
`gen_valid_hll_precision` generator so that the original "valid" bucket
is actually valid.