MAYBE?? Add a dummy `jmx` prop to schema
Cuttlefish will freakout and die and refuse to start a riak node if
there is an unknown/mapped property in riak.conf. I guess this is good
because...anyway, this is a speculative workaround for the upgrade from
riak_ee to riak+repl, since there is likely a `jmx` config setting in
riak.conf. Probably the right answer is to port riak_jmx to riak-2.2.5
and deprecate it.