Intermittent test failure fixes (#1657)
* Fix intermittent failure of `hll` bucket prop

The `gen_valid_bucket` call used regular `gen_hll_precision` meaning
that every now and again the _valid_ bucket had an _invalid_ hll
property. When this coincides with the update property also having the
same invalid hll property the resulting bucket has an invalid hll
property that matches the input and so it appears that validation
failed, and rightly the test fails. However validation worked, the
original bucket was invalid is all. This commit adds a
`gen_valid_hll_precision` generator so that the original "valid" bucket
is actually valid.

* Fix intermittent stat registration failure

A duplicate name was is used for the get_fsm_active stat, and also the
registration code is not robust. This commit removes the unused
duplicate alias and makes the registration code more robust by using
re-register.

* Remove vestiges of riak_search

* Tests failed when `make test` is run from Riak dir

Due to bad assumptions about paths to deps schema tests and cli tests
failed reliably from a top level checkout. This commit, while not pretty
or exact passes of both top level checkouts and local deps checkouts. I
don't think it is worth buring more time making it better or less
repetitive unless you do?