Add bucket property `node_confirms` for physical diversity (#1663)
* Add bucket property `node_confirms` for physical diversity

Briefly, some Riak deployments require that a write is on multiple
physical nodes before they're safe.

Given a good ring, using Primaries is one way to ensure this physical
diversity is to use `pw=2` (for example) to ensure two physical nodes
for the write. This has an availability cost. With 5 nodes and 2 nodes
down storing on 2 physical nodes is still very possible, but 2
primaries will leave a range of preflists unavailable to write. This
commit adds a bucket property and put option of `node_confrims` to the
declare the number of diverse physical node acks required for a write
to be successful. For the above example `node_confirms=2` would be
used, and as long as any two diverse nodes (primary OR fallback)
reply, the write is a success.

* Create Node-Diversity.md

Description and justification for new node_confirms option.

* Fewer arguments to put_core:init

When adding `node_confirms` a node confirms failure threshold was also
added, this brought the arity to 11 for put core init. All quorum
failure thresholds are calculated as `N+Q-1`. This commit moves that
calculation inside the put core. The arity of init is still 8, which is
high.

* Add `node_confirms` to PB API

objects/counters/datatypes all

* Merge typed defaults if prop not present

Typed buckets do not automatically inherit newly added default
properties (old-skool default buckets do thanks to the existence of
bucket fixups.)

As per the bug report, when PUTing to a Typed bucket after upgrade to
2.2.5 the call to riak_kv_util:expand_rw_value/4 throws if the
property `node_confirms` is not in the bucket properties, and for a
typed bucket created and activated before upgrade it won't be.

This is somewhat contentious. Ex-Basho @seancribbs has pointed out
that there is a reason for that throw, and that all nodes should agree
on the value, and misconfiguration can be an issue here.  There are
perhpas more complexities (we're retroactivley adding a reserved word,
for example.) Caveats in docs and probability should guard against
that latter though.

Given that the reasons not to do this are lost to time, and the
alternatives unpleasant, I'm content that this is a reasonable way for
typed buckets to inherit **NEW** properties added as defaults after
typed bucket creation.

There is an associated riak_core commit.

* Upstream rebar.config