Filter out head-only results before read_repair
Otherwise the current merge/2 code depends on reconcile/2, and this uses riak_object:remove_dominated/1.

remove_dominated/1 removes objects which are the same using the inherent de-duplication of sets:from_list/1.  However, if the objects differ by being a head or body response, these are no longer de-duplicated and a fake merge is required with unpredictable results.  e.g. the code could send an object with no contents as the repairing object!