Add rebuild of tictac support
Attempt to add support by pinging every 30 minutes and then checking in response to the ping to see if a rebuild is required.

Need to check that kv_index_tictactree will handle concurrent requests in case htis happens (e.g. if next_rebuild time is not moved on by ongoing request).

Perhaps need some backup to generate a poll should a send_after never be received (perhaps if mailbox is flushed without the vnode restarting?)