Spoof uplift of deleted objects
If we discover than an object is deleted in the HEAD request, we don't GET it (as there's nothing to GET).  However, the lack of non head_only responses confuses the final_action logic - and the conseqence of this is that whever we discover in a GET that the object is deleted everywhere, the action to reap the tombstone is not triggered.

Now when we discover an object is deleted, and it is a HEAD response, we spoof the ulift to a GET response by simply replacing the head_only value with anempty binary.