Track tictac rebuilds
Reverted previous commit to add tictac rebuilds, because tictac rebuilds were already going on!

This led to overlapping tictac rebuilds which caused crashes.  `kv_index_tictactree` makes preventing the overlaps the applictaion responsibility, so we now track some state in the riak_kv_vnode when we are rebuilding, to avoid rebuilding.

But what to do is a rebuild silently crashes, or goes on forever somehow.  We warn on this for now - we should probably restart the vnode?