Initial query definitions
Ignore code for now.  This commit is just to clarify intended scope of the feature.

The initial scope was just to provide the *_nval query definitions, as this is what is required for full_sync AAE using Tictac AAE.

The *_range query definitions are required to allow for AAE on a bucket/key_range.  This meets a need to perform full-sync AAE on a key range (e.g. timestamp-based keys in the past 24 hours) in a bucket.

The last two come for free with Tictac AAE, and have been long wished for ... as detecting the size of buckets, the existence of very large objects, and objects of siblings is veyr desirable.  This in many ways supercedes some of the intentions of kv_sweeper