Remove sample_size as an option
Previously I'd assumed I could do something through the plan callback.  However, this doesn't allow any modification to the coverage plan - it just allows for the ModState to be updated (to do somethig in process_results based on the coverage_plan).

don't wish to risk tampering with riak_core coverage_fsm for now, so won't allow the sampling in the query