Fix bug where `mbox_check` is set to false by client
the code in select_coordinator assumed that whenever
`mbox_check=false` it was set by a put fsm that had forwarded to the
executing put fsm, and therefore there would always be an entry in the
local preflist. HOWEVER, if the client were to use `mbox_check=false`
and the put fsm was started on a node not on the preflist, a crash
would occur. This change re-instates the old logic, with some
explanatory comments.