Initial gen_server code
Intention for the gen_server to on timeout pick the next item of the queue - and send a message to handle that event (when it is due).

When the gen_server receives the message, it should handle the event.  Then the reply from the exchange (or some timeout after assuming the exchange ahs crashed) should prompt a timeout - and extracting the next item from the queue.

Server should continue for ever in an eternal loop