Initial implementation of replrtq
First draft of underlying queue implementation based on riak_core_priority_queue.

this allows for a node to build a queue which can be a source of replication entries for a consumer on a sink cluster to fetch from.

The actual API to fetch is still required, as well as the sinkside implementation.