Initial draft of sink code
The sink code should allow for a src queue to be read from multipe peers on a src cluster.

There should be a fix number of workers, and the workers should exponentially back-off if there is no work to do.

Errors should lead to a peer not being used for a 60s period - to prevent all workers from being blocked by a single erroring peer.