Hook in real-time updates
Hook in real-time updates from PUT co-ordinators  ready for pipe-clean testing.

When a PUT coordinator is complete this is the first opportunity to repl - so it happens here and not after post-commit (via a hook).

To avoid fetching every recent PUT at the cost of each GET, there is a shortcut cache for recent PUTs kept at the coordinator, and the queue can indicate to fetch form that location.

Tombstones should be rpelicated whole, to avoid a slow repl sink from losing a race with tombstone reaping.

This does not work with w1c path (as with existing repl) - but extending this path to cast to the replrtq_src as well should be achievable.