Use state for temporary carrying of is_consistent
The if_none_match was tempoaraily overloaded with the atom consistent for consistent buckets.  However, it cannot have this type (as it is defined as a boolean, and the type cannot be changed in the auto-generated type file without amending the API).

Instead the process state will hold the is_consistent  information.  This allows for the GET to be skipped (at this stage) for  consistent PUTs, but the if_non_match still to be passed on as true.