GET_FSM with expected clock
If we have within a  GET_FSM an expectation of what the vector_clock should be (for example as we have fetched a reference for replication from the replication queue).  Then we can start a get_core with a reference to that clock - and the GET_FSM will exit to provide a response as soon as that expected clock is encountered (and otherwise wait for n responses).

This prevents a race condition which may lead to replication events being missed when the repl_cache is not enabled.

Further details - https://github.com/basho/riak_kv/issues/1691#issuecomment-507582225

This will cause replication failures during intermittent node failures.  However, such failures can be avoised by turning on the repl_cache (whereby this path will be bypassed).  That should remain as the preferred option.