Finding best answers refactor
Adding further unit tests demonstrated some non-determinism in the find_bestobject/1 function in that it might consider different answers to be siblings depending on the order the answers would be received.  That is to say dominated siblings may or may not be included.

It would be better to make this predictable, so that any riak_test that proves behaviour with siblings can't be dependent on the order of answers - increasing confidence that such situations are handled correctly