Clarify tombstone handling with heads
When a head is a tombstone, the get_core will convert it into something object like (using riak_object:spoof_deletedobject/1), when the response is added to the get_core.  Hence we don't need to consider when running find_bestobject that head objects could be deleted.

There had previously been logic in merge_heads to discount tombstone head responses as things that need to eb fetched - but they shouldn't be head-like things at that stage due to the use of spoof_deletedobjects.  This logic is therefore superfluous.

Also contains a check that the riak_kv_util:is_x_deleted/1 function does work as expected on HEAD-like responses.