Return fail from prepare_put on unexpected events
Currently certain unxpected events in prepare_put will cause the vnode to crash.  It would be preferable to fail the PUT, but also log something out to offer a clue as to how these unexpected events are occurring.

Ther eis a difficult balance here.  Letting it crash prudces the crash log, which contains additional information about the request and the vnode state.  However, this hasn't proven to be enough to resolve why this is occurring - so better now to gather some further information.  Also, it is not clear if this issue could occurr with sufficient frequency to cause further problems due to continuous restart of the vnode.

As the problem can't be recreated, the code branch is not tested.  To provide some assurance it won't crash anyway, spec has been added to allow dialyzer to review.