Addition of erase folds
This adds a riak_kv_eraser with equivalent behaviour to the riak_kv_reaper (although those similarities are not currently captured via a behaviour).  The eraser can queue up deletes, and there is an addition of an aae_fold to send deletes to the queue.

Some extra work is still required:

1. Make this a behaviour
2. Add the API changes for fold
3. Should the eraser still doubl-check the vector clock
4, Should it like the reaper only erase when primaries are up