Remove repl_cache
Originally it was assumed it would improve performance of replication if each vnode kept an ets-backed cache of recently co-ordinated PUTs - and objects on the repl queue could reference these vnodes for fetching via special fetch request which would check the cache.

In real world performance tests, this improved write throughput slightly when under heavy write pressure and not disk I/o bound - but in the general case it increase tail latency for replication (and hence mean latency).  More consistent behaviour can be achieved with the repl_cache disabled.

This commit removes the repl_cache on a YAGNI principle.