Change sink worker to allow for multiple queues in config
Peer configuration can now contain a queue-name, and hence we can have mutliple queues in the configuration.

Also allows for the worker count to be changed for a given queue.

There is now an iteration number kept - to make it easier when switching configurations to discard in-flight work items that relate to old configuration.

This change means that protocol no longer defaults to 'http' - it must be explicity set.