Ping the aae_controller
This is equivalent to how legacy AAE would regularly do a sync not async PUT to ensure that the AAE database is not getting behind.

Wanted to change though so that there is no risk of dealocks, and there is improved monitoring around what is happening when the AAE store is behind.