Allow reduce functions on index keydata
The index term can be added to keydata, and hence reduce functions can be applied that extract and filter from overloaded index terms.

A map can be added, but immediately prior to the map, the object would be fetched.

This makes it a bit articicial as some of these functions should be map not reduce (as they don't reduce).  however, it still works as the functions pass the commutative, associative and idempotent tests (with trivial test cases).