Allow prereduce to be forced
Will only work if first reduce.

Simplest change to allow first reduce of an index map/reduce to be a prereduce - previously the code only allows prereduce when a reduce follows a map, and a prereduce is requested as a property of the query.

This will lead to failures if used outside of index map/reduce queries, or if this is not the first reduce statement in the query, or in some other unexpected combination.