Switch to using fetch_clocks_range
This also appears to be much quicker within `riak_test`.  Why, is not clear.  Perhaps related to queueing at aae_runner as perfomance log within aae_runner is fast even when fetch_clocks_nval returns a slow response.

This change is leading towards the possibility of adding modified_range to fetch_clocks_nval, and so allowing for hourcheck and daycheck as well as allcheck in ttaaefs schedules.