Add configurable pause
There exists a race condition here, and wan to try and take the minimal step to mitigate this condition.  There is not hurry at this point, as the client has already been informed of the success/failure of the delete.

Assumption is that a pause approximately equal to the intra-cluster network delay will be sufficient.

Note that the race can be cleared up even if it happens (e.g. by repl or by manual reap).  Aim is to contorl the volume at which it happens (to make it easier for repl to clear up without intervention of manual repl).