Add support for range_check
Replaces the boost concept.  Now range_checks can be added to the schedule, and when a delta is discovered the range of the delta will be set to be used in the next range_check query.

If there is no range set, then the range_check is a null op.  The range is always unset before each range_check, and will only be reset if sufficient results are found.

The expectation is that fewer hour/day/all checks should need to be configured as the bulk of the work will be managed by range_check.  As the range_check is handling an identified problem - it will be run with boosted max_results.