Add scheduling support for range_sync
Also expose the set/get/clear functions for the range - make it easier to control this from a remote_console and/or tests.

Need to account for the fact that rnage queries may now be requested on NVal tests - and so NVal may be an integer not range on fetch_clocks_range.