ttaaefs_manager - not too clever
Trying to pick a single bucket more often was useful in some tests, however it could lead to strange results as the "top" bucket drained of results (especially if mroe than one range_check picked up the last gasp.  This led to more lumpy resolution.

The repairing behaviour has a big impact on performance, and cna be turned on if required.  Leave repair to intra-cluster AAE.  don't duplicate this effort.