BE AWARE - potential breaking change
Tombstones do not currently have a last modified date.  This has not caused any issues, as last modified dates are only checked:

https://github.com/basho/riak_kv/blame/d17495741ed001dcb54fa9e75e0dff026baaefde/src/riak_kv_vnode.erl#L3048-L3049

This would have previously blown up before the relative safe get_last_modified function was added, as opposed to doing a dict:fetch for the tag (which would have triggered an exception).

This is necessary to have time-based rpelication work correctly.  Currently there can be issues where recent tombstone discrepancies cna only be picked up by all_check (even if they are recent).  This is confusing.

Should we just live with the confusion rather than risk changing this?  But it feels like tomstones not having a LMD is just a mistake i.e. it being useful is just a scenario which was missed, and there are no issues.

Note, as a side effect, tombstones also get a vtag.

Perhaps some client somwhere will find this weird/unexpected?