Prioritise changes from 3.0.2
Changes for 2.9.8 release, take some of the priority fixes in 3.0.2, but don't include all of the new full-sync optimisation functionality - in particlaur the hour_check, day_check, range_check feature for nval replication.  Also, the adding of. a last modified date for tombstones.

It is recommended that 3.0.2 should be used if ttaaefs_fullsync is to be enabled for non-trivial key-store sizes where non-trivial deltas may occur, and manual intervention (e.g. for range_repl) might not be possible..