Updated the second example
The `extract_buckets` primitive takes a map. This map can either work on ranges, or just values or possibly later regexp etc, like the filters. This is not clearly written down and for the moment I define the `extract_buckets` to always work on a map of ranges. Some thought need to go into how to represent the different alternatives