max integer()/undefined -> undefined
As a safety check the definition of CSP checks it as at least 1 to avoid badarith when divided.  However, when called from eqc tests, this property will be undefined (should be OK when started via Riak), and max(undefined, 1) is undefined.  So this needs a default to work safely from tests