Mas ghafix (#1785)
* Update erlang.yml

...

* max integer()/undefined -> undefined

As a safety check the definition of CSP checks it as at least 1 to avoid badarith when divided.  However, when called from eqc tests, this property will be undefined (should be OK when started via Riak), and max(undefined, 1) is undefined.  So this needs a default to work safely from tests

* Recommended fixes

Following review from Thomas

* Revert to develop-3.0 restriction

Not sure what the best setting is, but for now reverted to make consistent with other GHA settings in the riak environment