Add config for worker_pool_size
Previously only possible to set this via advanced.config.

The default has been reduced, as with a large pool, it is possible that context switching between a vast number of concurrent folds could be slower overall than simply queueing some folds to allow others to complete.

Even on volume tests with 1% of transactions using 2i queries, a pool size of 4 is never exhausted.