Change time out checking in unit test
Previously this relied on catching a reply to a sync_send_event which had timed out through a receive.  This appears not to work now - a receive for the same PID as the crashed send, will not receive a response.  It is assumed this is related to smarter references for sync replies in OTP 24.

A more natural way is used now (albeit one that is dependent on timers).  We allow the sync request to get a delayed response, and use a timer to confirm it is delayed and not immediate.