Correctly encode clocks pb/http
The encoding of vclocks differs between http and pb due to the need to base64 encode for http.

Case clauses change for auto_check, as otherwise may have requested a range_check without a range or previous success, when drop_next_auto_check() was true, even when the peer queue is not diabled.