Refinements after further volume testing
The coordination of syncs between clusters caused performance spikes in testing.  A simple (and optional) way round this is the configuration of cluster slice numbers.

Also, this change simplifies whether or not a range will be defined.

As auto_check used by default - other checks are just a "finder", and so can be accelerated by using less max_results.  Even with a relatively small sample size, significant "clipping" of the boundary is unlikely (e.g. only 3% chance the lmd range will be clipped by more than 10%).