Potential for broken trees
If the aae_trees are broken on multiple nodes - i.e. the same segment represented incorrectly in the same way on all 3 nodes, then internal anti-entropy cannot repair this.

This can lead to a situation where inter-cluster anti-entropy can never resolve.  This can be resolved by enforcing the rebuild of the trees within the aae_fold fetch clocks (although this has an efficiency overhead).

Rather than rely on operator intervention to trigger in this case, this code change will spot {clock_compare,0} on an all_check  and set this parameter on the node.  Eventually, if most nodes have peers, then this will be resolved a sit is turned on node by node.

Once a sync is successful, it will be turned back off.