Catch object corruption in rebuild folds
Don't add a corrupted object when rebuilding - instead log, skip past, and prompt read repair.

The behaviour before was to avoid crashing on inserting a corrupted object (whereby the vector clock cannot be extracted for hashing).  Not clear why this is still inserted.  This behaviour is maintained, with the alternative behaviour during the rebuild.

The same handle_corrupted_object/4 fun is then re-used during parallel mode tictac aae rebuilds, with a new version of kv_index_tictactree which catches exceptions and allows an external function to be used for handling them).