Schema backwards compatability
The replrtq_srcqueuelimit was removed, as the queue is now an overflow queue and should have its limit configured via replrtq_overflow_limit.

However, removing the old config item altogether will result in upgrades not being possible without changes to configuration files - so the old option is instead retained as a hidden config item to be ignored.