Mas i1807 overflowqueues (#1809)
* Bound reap/erase queues

Also don't log queues on a crash - avoid over-sized crash dumps

* Create generic behaviour for eraser/reaper

Have the queue backed to disk, so that beyond a certain size it overflows from memory to disk (and once the on disk part is consumed from the queue the files are removed).

* Stop cleaning folder

Risk of misconfiguration leading to wiping of wrong data.  Also starting a job may lead to the main process having its disk_log wiped.

* Setup folders correctly

Avoid enoent errors

* Correct log

* Further log correction

* Correct cleaning of directories for test

* Switch to action/2

* Update eqc tests for refactoring of reaper/eraser

* Improve comments/API

* Pause reaper on overload

Check for a soft overload on any vnode before reaping.  This will add some delay - but reap has been show to potentially overload a cluster ... availability is more important than reap speed.

There is no catch for {error, mailbox_overload} should it occur - it should not as the mailbox check should prevent it.  If it does, the reaper will crash (and restart without any reaps) - return to a safe known position.

* Adjustments following review

The queue file generated, are still in UUID format, but the id now incorporates creation date.  This should make it easier to detect and clean any garbage that might be accrued.

One use case where '++' is used has also been removed (although a lists:flatten/1 was still required at the end in this case)
# Conflicts:
#	priv/riak_kv.schema
#	rebar.config
#	src/riak_kv_overflow_queue.erl
#	src/riak_kv_reaper.erl
#	src/riak_kv_replrtq_src.erl
#	src/riak_kv_test_util.erl