Handle handoff put through standard put code
Rather than replicate piecemeal the standard PUT code within do_diffobj_put/3, instead use do_handoff_put/3 and use standard prepare_put/2 and perform_put/3 functions used in normal PUTs.

The effect of this is that any optimisations in the normal PUT workflow, will now automatically be used for handoffs.  Of particular relevance at this point is the HEAD (not GET) before PUT optimisation available with leveled backend.  If there are large objects, and objects which already exist in the receiving vnode are to be handed off (such as in hinted handoff), then this increases efficiency.

Some spec improvements to help with some editors that do not like fun() type.  Some indent reductions to improve readability.