Mas i1847 putapi (#1848)
* Never GET before PUT

As the if_not_modified and if_none_match are not supported via the HTTP API for non-consistent PUT - simply build object from a new object, as with PB API.

* More webmachine friendly override

resource_exists/2 is a key part of the flow for both GET and PUT, as conditional HTTP headers require this check for PUTs.  therefore only override resource_exists (and don't fetch) when it is a PUT, and those conditional headers do not exist.

* Attempt to tidy and refactor delete

So that delete does not require a fetch

* Pipe-cleaned delete path

* Add if_not_modified conflict check

To mimic if_not_modified feature via PB API

* Use hyphen not underscore

To be consistent with other HTTP headers

* Revert to 404 on if DELETE not_found

Also ensure the timeouts passed in in a delete is respected, and passed through the riak_client to the FSM.