Log Fragmentation (#1852)
* Log Fragmentation

With the leveled backend, memory fragmentation is an ongoing concern.  So, here the regular compaction callback is used to log information about carrier sizes and carrier block sizes for key allocators.

* Correct log format

* Update riak_kv_leveled_backend.erl

* Update to account for review comments

* Spawn function to log

There may be a very large number of allocators, so the cost of calling recon_alloc:fragmentation/1 and processing the outputs could be high - so don't lock the vnode in this case.

* Don't make log function depend on format of recon_alloc:fragmentation/1

* Update rebar.config