HTTP client may respond after timeout
The ibrowse http client  will gen_server:call on send_req with a timeout, and catch any timeout error.

This means that the client may receive an {error, reqd_timeout}, but the gen_server has not crashed and so may later send a gen_server:reply to the riak_kv_replrtq_peer.

These errors need to be picked up in handle_info to avoid noise of repeated crashes.

This is not a problem in riak_kv_replrtq_snk, as in the process we are always using the http client from within a short-lived spawned process.  Likewise in riak_kv_ttaaefs_manager the client is sued within a short-lived aae_exchange.